Добавляет BoardController и переносит функцию рендера доски
